week-13-upload-to-s3-vultr/requirements.txt
boto3
logging
python-dotenv
